Remove exercises from intro to hooks
