--- src/pages/Dashboard/Dashboard.example-3.tsx
@@ -1,30 +1,0 @@
-import React, { useContext, useEffect } from 'react';
-import { Redirect } from 'react-router';
-import Paper from 'src/common/components/Paper';
-import Row from 'src/common/components/Row';
-import ViewerContext from 'src/common/components/ViewerContext';
-
-const Dashboard: React.FunctionComponent = () => {
-  const { viewer } = useContext(ViewerContext);
-
-  useEffect(() => {
-    alert('Welcome to dashboard!');
-
-    return () => {
-      alert('Navigating away from dashboard');
-    };
-  });
-
-  if (!viewer) {
-    return <Redirect to="/" />;
-  }
-
-  return (
-    <Paper>
-      <h1>Dashboard</h1>
-      <Row>Welcome {viewer.email}!</Row>
-    </Paper>
-  );
-};
-
-export default Dashboard;
