+++ src/pages/Dashboard/Dashboard.test.tsx
@@ -1,24 +1,9 @@
-/**
- * `useEffects` tests is a bit harder to test.
- * The function passed to useEffect fires after layout and paint, during a deferred event.
- * This means in jest, we cannot easily capture when it's fired
- *
- * Reference:
- * https://reactjs.org/docs/hooks-reference.html#timing-of-effects
- *
- * See discussion:
- * https://github.com/kentcdodds/react-testing-library/issues/215
- *
- * Sample implementation:
- * https://github.com/threepointone/react-act-examples/blob/master/act-examples.test.js
- */
-
 const ReactDOM = require('react-dom');
 import React from 'react';
 import { StaticRouter } from 'react-router';
 import { UsersProvider } from 'src/common/components/UsersContext/UsersContext';
 import { ViewerProvider } from 'src/common/components/ViewerContext/ViewerContext';
-import Dashboard from './Dashboard.example-3';
+import Dashboard from './Dashboard';
 
 const { act } = require('react-dom/test-utils');
