+++ src/pages/Dashboard/index.tsx
@@ -1,2 +1,1 @@
 export { default } from './Dashboard';
-// export { default } from './Dashboard.example-3';
