+++ src/pages/Login/index.tsx
@@ -1,3 +1,1 @@
 export { default } from './Login';
-// export { default } from './Login.answer-1';
-// export { default } from './Login.answer-2';
