+++ src/pages/Logout/index.tsx
@@ -1,2 +1,1 @@
 export { default } from './Logout';
-// export { default } from './Logout.answer-3';
