+++ src/pages/Signup/index.tsx
@@ -1,3 +1,1 @@
 export { default } from './Signup';
-// export { default } from './Signup.example-1';
-// export { default } from './Signup.example-2';
